created reusable SEO component
